+++ example/lib/src/my_mixin.dart
@@ -1,0 +1,5 @@
+mixin MyMixin {
+  void foo() {
+    //nothing really
+  }
+}