+++ lib/src/generator/entity_generator.dart
@@ -12,7 +12,8 @@ class EntityGenerator {
         name: reqAttrValue(node, "name"),
         copyWith: asBool(optAttrValue(node, "copyWith", "true")),
         generateEntity: asBool(optAttrValue(node, "generateEntity", "true")),
-        apiProxies: asSplitList(attrValue(node, "apiProxy")));
+        apiProxies: asSplitList(attrValue(node, "apiProxy")),
+        mixins: asSplitList(attrValue(node, "mixin")));
     final serializers = attrValue(node, "serializers");
     if (serializers != null && serializers.trim().isNotEmpty) {
       result.serializers.addAll(asSplitList(serializers));
@@ -100,8 +101,7 @@ class EntityGenerator {
       default:
         {
           final xmlName = node.name.toString();
-          final rawName =
-              isOptional ? xmlName.substring("optional".length) : xmlName;
+          final rawName = isOptional ? xmlName.substring("optional".length) : xmlName;
           var varType = rawName == "dynamic" ? "dynamic" : rawName.pascalCase;
           final expects = asSplitList(attrValue(node, "expectOnly"));
           result = Field(
@@ -128,9 +128,7 @@ class EntityGenerator {
     var isPlain = asBool(optAttrValue(node, "plain", "true"));
     final expects = asSplitList(attrValue(node, "expectOnly"));
     final innerType = reqAttrValue(node, "innerType");
-    if (expects.isNotEmpty ||
-        innerType == "dynamic" ||
-        !Field.innerDartTypes.contains(innerType)) {
+    if (expects.isNotEmpty || innerType == "dynamic" || !Field.innerDartTypes.contains(innerType)) {
       isPlain = false;
     }
     return Field.list(
@@ -152,9 +150,7 @@ class EntityGenerator {
     var isPlain = asBool(optAttrValue(node, "plain", "true"));
     final expects = asSplitList(attrValue(node, "expectOnly"));
     final innerType = reqAttrValue(node, "valueType");
-    if (expects.isNotEmpty ||
-        innerType == "dynamic" ||
-        !Field.innerDartTypes.contains(innerType)) {
+    if (expects.isNotEmpty || innerType == "dynamic" || !Field.innerDartTypes.contains(innerType)) {
       isPlain = false;
     }
     return Field.map(