+++ lib/src/model/entity.dart
@@ -8,10 +8,12 @@ class Entity {
   final List<Field> fields = [];
   final List<String> serializers = [];
   final List<String> apiProxies;
+  final List<String> mixins;
 
   Entity(
       {required this.name,
       required this.copyWith,
       required this.generateEntity,
-      required this.apiProxies});
+      required this.apiProxies,
+      required this.mixins});
 }